+++ apps/drain-safe/src/GlobalStyle.ts
@@ -1,16 +1,17 @@
 import { createGlobalStyle } from 'styled-components'
-import avertaFont from '@gnosis.pm/safe-react-components/dist/fonts/averta-normal.woff2'
-import avertaBoldFont from '@gnosis.pm/safe-react-components/dist/fonts/averta-bold.woff2'
 
 const GlobalStyle = createGlobalStyle`
     html {
         height: 100%
+        font-family: 'DM Sans', sans-serif;
     }
 
     body {
        height: 100%;
        margin: 0px;
        padding: 0px;
+       background-color: #f6f6f6;
+       font-family: 'DM Sans', sans-serif;
     }
 
     #root {
@@ -23,12 +24,6 @@ const GlobalStyle = createGlobalStyle`
         width: 100% !important;
     }
 
-    @font-face {
-        font-family: 'Averta';
-        src: local('Averta'), local('Averta Bold'),
-        url(${avertaFont}) format('woff2'),
-        url(${avertaBoldFont}) format('woff');
-    }
 `
 
 export default GlobalStyle