+++ apps/siwe-delegate-manager/src/GlobalStyles.tsx
@@ -1,0 +1,11 @@
+import { createGlobalStyle } from 'styled-components'
+
+const GlobalStyles = createGlobalStyle`
+  body {
+    height: 100%;
+    margin: 0px;
+    padding: 0px;
+  }
+`
+
+export default GlobalStyles