+++ apps/tx-builder/src/components/forms/validations/basicSolidityValidation.ts
@@ -3,6 +3,7 @@ import abiCoder, { AbiCoder } from 'web3-eth-abi'
 
 import { parseInputValue } from '../../../utils'
 import { NON_SOLIDITY_TYPES } from '../fields/fields'
+import { isEthersError } from '../../../typings/errors'
 
 const basicSolidityValidation = (value: string, fieldType: string): ValidateResult => {
   const isSolidityFieldType = !NON_SOLIDITY_TYPES.includes(fieldType)
@@ -11,8 +12,18 @@ const basicSolidityValidation = (value: string, fieldType: string): ValidateResu
       const cleanValue = parseInputValue(fieldType, value)
       const abi = abiCoder as unknown // a bug in the web3-eth-abi types
       ;(abi as AbiCoder).encodeParameter(fieldType, cleanValue)
-    } catch (error: any) {
-      return `format error. details: ${error.reason || error.toString()}`
+    } catch (error: unknown) {
+      let errorMessage = error?.toString()
+
+      const errorFromEthers = isEthersError(error)
+      if (errorFromEthers) {
+        if (error.reason.toLowerCase().includes('overflow')) {
+          return 'Overflow error. Please encode all numbers as strings'
+        }
+        errorMessage = error.reason
+      }
+
+      return `format error. details: ${errorMessage}`
     }
   }
 }